+++ CHANGELOG.md
@@ -1,1 +1,4 @@
 autowrap 0.22.9
+
+- Typing of a majority of classes and functions of the codebase
+- Support for "const Members" as alternative to "wrap-constant" (experimental)